testing/web-platform/tests/atomics/window-and-any-worker/wait/negative-timeout/negative-timeout.html

<!doctype html>
<meta charset=utf-8>
<title></title>

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>

<body>
  <iframe id="myiframe" style="display: none" src="../empty.html" ></iframe>
</body>

<script>

// Copyright (C) 2015 Mozilla Corporation.  All rights reserved.
// This code is governed by the BSD license found in the LICENSE file.

/*---
description: >
  Test that Atomics.wait times out with a negative timeout
---*/

window.addEventListener("load", function(e) {
  // Window->Worker.
  async_test(function(t) {
    let worker = new Worker('wait-on-zero-negative.js');
    worker.onmessage = this.step_func_done(function(e) {
      assert_equals(e.data, "timed-out");
    });
    let ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
    worker.postMessage([ia.buffer]);
  }, "atomics-dedicated-worker-wait-negative-timeout");
  // Window->Worker->Worker.
  async_test(function(t) {
    let worker = new Worker('negative-timeout-worker.js');
    worker.onmessage = this.step_func_done(function(e) {
      assert_equals(e.data, "timed-out");
    });
    worker.postMessage([]);
  }, "atomics-dedicated-worker-wait-negative-timeout-from-worker")
  // Window->SharedWorker->Worker.
  async_test(function(t) {
    let worker = new SharedWorker('negative-timeout-shared-worker.js');
    worker.port.start();
    worker.port.onmessage = this.step_func_done(function(e) {
      assert_equals(e.data, "timed-out");
    });
    worker.port.postMessage([]);
  }, "atomics-dedicated-worker-wait-negative-timeout-from-shared-worker")
  // Window->IFrame->Worker.
  async_test(function(t) {
    let iframe = document.getElementById('myiframe');
    let contentWindow = iframe.contentWindow;
    let test = this;
    contentWindow.onmessage = function(e) {
      let worker = new Worker('wait-on-zero-negative.js');
      worker.onmessage = test.step_func_done(function(e) {
        assert_equals(e.data, "timed-out");
      });
      let ia = new Int32Array(new SharedArrayBuffer(Int32Array.BYTES_PER_ELEMENT));
      worker.postMessage([ia.buffer]);
    }
    contentWindow.postMessage([], '*');
  }, "atomics-dedicated-worker-wait-negative-timeout-from-iframe")
})

</script>
